+++ backend/src/main/java/com/prograngers/backend/repository/solution/SolutionRepository.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.repository.solution;
 
-import com.prograngers.backend.entity.Member;
+import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.entity.Solution;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
