+++ backend/src/test/java/com/prograngers/backend/fixture/CommentFixture.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.fixture;
 
 import com.prograngers.backend.entity.Comment;
-import com.prograngers.backend.entity.Member;
+import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.entity.Solution;
 import lombok.AllArgsConstructor;
