+++ backend/src/test/java/com/prograngers/backend/fixture/MemberFixture.java
@@ -1,14 +1,7 @@
 package com.prograngers.backend.fixture;
 
-import com.prograngers.backend.entity.Member;
-import com.prograngers.backend.entity.Problem;
-import com.prograngers.backend.entity.constants.JudgeConstant;
+import com.prograngers.backend.entity.member.Member;
 import lombok.AllArgsConstructor;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-
-import static com.prograngers.backend.entity.constants.JudgeConstant.백준;
 
 @AllArgsConstructor
 public enum MemberFixture {
